--- lib/lib/screens/settings_screen.dart
@@ -1,1 +1,0 @@
-// TODO Implement this library.
