--- lib/widgets/audio_player_widget.dart
@@ -1,238 +1,0 @@
-import 'dart:io';
-import 'package:flutter/material.dart';
-import 'package:just_audio/just_audio.dart';
-import 'package:provider/provider.dart';
-import 'package:sonify/models/theme_provider.dart';
-import 'package:sonify/utils/theme_config.dart';
-
-class AudioPlayerWidget extends StatefulWidget {
-  final String audioPath;
-  final Function? onSave;
-
-  const AudioPlayerWidget({
-    Key? key,
-    required this.audioPath,
-    this.onSave,
-  }) : super(key: key);
-
-  @override
-  State<AudioPlayerWidget> createState() => _AudioPlayerWidgetState();
-}
-
-class _AudioPlayerWidgetState extends State<AudioPlayerWidget> {
-  late AudioPlayer _audioPlayer;
-  bool _isPlaying = false;
-  bool _isLoading = true;
-  Duration _duration = Duration.zero;
-  Duration _position = Duration.zero;
-
-  @override
-  void initState() {
-    super.initState();
-    _initializePlayer();
-  }
-
-  @override
-  void didUpdateWidget(AudioPlayerWidget oldWidget) {
-    super.didUpdateWidget(oldWidget);
-    if (oldWidget.audioPath != widget.audioPath) {
-      _disposePlayer();
-      _initializePlayer();
-    }
-  }
-
-  @override
-  void dispose() {
-    _disposePlayer();
-    super.dispose();
-  }
-
-  void _disposePlayer() {
-    _audioPlayer.dispose();
-  }
-
-  Future<void> _initializePlayer() async {
-    _audioPlayer = AudioPlayer();
-
-    try {
-      setState(() {
-        _isLoading = true;
-        _position = Duration.zero;
-      });
-
-      // Check if file exists
-      final file = File(widget.audioPath);
-      if (!await file.exists()) {
-        throw Exception('Audio file not found');
-      }
-
-      // Set audio source
-      await _audioPlayer.setFilePath(widget.audioPath);
-
-      // Get duration
-      final duration = await _audioPlayer.duration;
-      setState(() {
-        _duration = duration ?? Duration.zero;
-      });
-
-      // Listen to player state changes
-      _audioPlayer.playerStateStream.listen((state) {
-        if (state.processingState == ProcessingState.completed) {
-          setState(() {
-            _isPlaying = false;
-            _position = _duration;
-          });
-        }
-      });
-
-      // Listen to position changes
-      _audioPlayer.positionStream.listen((position) {
-        setState(() {
-          _position = position;
-        });
-      });
-    } catch (e) {
-      debugPrint('Error initializing audio player: $e');
-    } finally {
-      setState(() {
-        _isLoading = false;
-      });
-    }
-  }
-
-  String _formatDuration(Duration duration) {
-    String twoDigits(int n) => n.toString().padLeft(2, '0');
-    final minutes = twoDigits(duration.inMinutes.remainder(60));
-    final seconds = twoDigits(duration.inSeconds.remainder(60));
-    return '$minutes:$seconds';
-  }
-
-  Future<void> _togglePlayback() async {
-    if (_isPlaying) {
-      await _audioPlayer.pause();
-    } else {
-      await _audioPlayer.play();
-    }
-
-    setState(() {
-      _isPlaying = !_isPlaying;
-    });
-  }
-
-  Future<void> _seekTo(double position) async {
-    final newPosition = Duration(milliseconds: (position * _duration.inMilliseconds).round());
-    await _audioPlayer.seek(newPosition);
-    setState(() {
-      _position = newPosition;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final isDarkMode = Provider.of<ThemeProvider>(context).isDarkMode;
-    final accentColor = isDarkMode ? darkAccentColor : lightAccentColor;
-    final primaryColor = isDarkMode ? darkPrimaryColor : lightPrimaryColor;
-
-    return Column(
-      children: [
-        Row(
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-          children: [
-            Text(
-              _formatDuration(_position),
-              style: TextStyle(
-                color: isDarkMode ? Colors.white70 : Colors.black54,
-              ),
-            ),
-            Text(
-              _formatDuration(_duration),
-              style: TextStyle(
-                color: isDarkMode ? Colors.white70 : Colors.black54,
-              ),
-            ),
-          ],
-        ),
-        const SizedBox(height: 8),
-        Row(
-          children: [
-            // Play/Pause button
-            InkWell(
-              onTap: _isLoading ? null : _togglePlayback,
-              child: Container(
-                width: 48,
-                height: 48,
-                decoration: BoxDecoration(
-                  shape: BoxShape.circle,
-                  color: primaryColor,
-                ),
-                child: _isLoading
-                    ? const Padding(
-                  padding: EdgeInsets.all(12.0),
-                  child: CircularProgressIndicator(
-                    strokeWidth: 2,
-                    color: Colors.white,
-                  ),
-                )
-                    : Icon(
-                  _isPlaying ? Icons.pause : Icons.play_arrow,
-                  color: Colors.white,
-                ),
-              ),
-            ),
-            const SizedBox(width: 12),
-            // Seek bar
-            Expanded(
-              child: SliderTheme(
-                data: SliderThemeData(
-                  trackHeight: 4,
-                  thumbShape: const RoundSliderThumbShape(enabledThumbRadius: 8),
-                  overlayShape: const RoundSliderOverlayShape(overlayRadius: 16),
-                  activeTrackColor: accentColor,
-                  inactiveTrackColor: isDarkMode ? Colors.grey[800] : Colors.grey[300],
-                  thumbColor: accentColor,
-                  overlayColor: accentColor.withOpacity(0.2),
-                ),
-                child: Slider(
-                  value: _position.inMilliseconds / (_duration.inMilliseconds == 0 ? 1 : _duration.inMilliseconds),
-                  onChanged: _isLoading ? null : _seekTo,
-                ),
-              ),
-            ),
-            // Save button (if provided)
-            if (widget.onSave != null) ...[
-              const SizedBox(width: 12),
-              InkWell(
-                onTap: () => widget.onSave!(),
-                child: Container(
-                  padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 6),
-                  decoration: BoxDecoration(
-                    color: accentColor,
-                    borderRadius: BorderRadius.circular(16),
-                  ),
-                  child: Row(
-                    mainAxisSize: MainAxisSize.min,
-                    children: [
-                      Icon(
-                        Icons.save,
-                        color: isDarkMode ? Colors.black : Colors.white,
-                        size: 16,
-                      ),
-                      const SizedBox(width: 4),
-                      Text(
-                        'Save',
-                        style: TextStyle(
-                          color: isDarkMode ? Colors.black : Colors.white,
-                          fontWeight: FontWeight.bold,
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-            ],
-          ],
-        ),
-      ],
-    );
-  }
-}
